requirements.txt
numpy==1.18.2
torch==1.4.0
torchvision==0.5.0
matplotlib==3.2.1
tensorflow-gpu
tensorboard
terminaltables==3.1.0
pillow==6.2.1
tqdm